# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Research Lab API
  description: |
    Enterprise-grade API for managing LLM research experiments, model evaluations,
    datasets, and prompt templates.

    ## Authentication

    This API supports two authentication methods:

    1. **JWT Bearer Token**: For user-based authentication
       - Obtain tokens via the `/auth/login` endpoint
       - Include in requests: `Authorization: Bearer <token>`

    2. **API Key**: For service-to-service authentication
       - Include in requests: `X-API-Key: <api_key>`

    ## Rate Limiting

    API requests are rate-limited based on authentication tier:
    - **Free Tier**: 100 requests/minute
    - **Standard Tier**: 1,000 requests/minute
    - **Premium Tier**: 10,000 requests/minute
    - **Enterprise Tier**: Unlimited

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

    ## Pagination

    List endpoints support pagination via query parameters:
    - `page`: Page number (1-indexed, default: 1)
    - `page_size`: Items per page (default: 20, max: 100)
    - `cursor`: Cursor for cursor-based pagination

    ## Error Handling

    Errors are returned in a consistent format:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Invalid request parameters",
...
        ]
      }
...
servers:
...
paths:
...
  /datasets:
    post:
      summary: Create a new dataset
      operationId: createDataset
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatasetRequest'
      responses:
        '201':
          description: Dataset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      summary: List all datasets
      operationId: listDatasets
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of datasets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDatasetResponse'
  /datasets/{id}:
    get:
      summary: Get dataset by ID
      operationId: getDataset
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: Dataset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update dataset
      operationId: updateDataset
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDatasetRequest'
      responses:
        '200':
          description: Dataset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete dataset
      operationId: deleteDataset
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '204':
          description: Dataset deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{id}/versions:
    post:
      summary: Create a new dataset version
      operationId: createDatasetVersion
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatasetVersionRequest'
      responses:
        '201':
          description: Version created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetVersionResponse'

    get:
      summary: List dataset versions
      operationId: listDatasetVersions
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetVersionResponse'
  /datasets/{id}/upload:
    post:
      summary: Get presigned upload URL
      operationId: getUploadUrl
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
  /datasets/{id}/download:
    get:
      summary: Get presigned download URL
      operationId: getDownloadUrl
      tags: [datasets]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'

  # ==================== PROMPTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service accounts
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DatasetId:
      name: id
      in: path
      required: true
      description: Dataset ID
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "NOT_FOUND"
            message:
              type: string
              example: "Resource not found"
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request parameters"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean

    # ==================== EXPERIMENTS ====================
...
    CreateDatasetRequest:
      type: object
      required: [name, s3_path, schema]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        s3_path:
          type: string
        schema:
          type: object
    UpdateDatasetRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        schema:
          type: object
    CreateDatasetVersionRequest:
      type: object
      required: [s3_path]
      properties:
        s3_path:
          type: string
        schema:
          type: object
    DatasetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        s3_path:
          type: string
        sample_count:
          type: integer
          format: int64
        schema:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
    DatasetVersionResponse:
      type: object
      properties:
        version:
          type: integer
        dataset_id:
          type: string
          format: uuid
        s3_path:
          type: string
        sample_count:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    PaginatedDatasetResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DatasetResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    UploadUrlResponse:
      type: object
      properties:
        upload_url:
          type: string
          format: uri
        s3_path:
          type: string
    DownloadUrlResponse:
      type: object
      properties:
        download_url:
          type: string
          format: uri

    # ==================== PROMPTS ====================
...